fix(gpu-zk): address review findings for Phase 3/4
- Widen point_offset + n to uint64_t to prevent u32 overflow in bounds check
- Replace `as u32` casts with try_from().expect() per CLAUDE.md rules

+++ backends/zk-cuda-backend/src/c_wrapper.cu
@@ -507,7 +507,7 @@ void zk_g2_msm_cached_async(
     PANIC_IF_FALSE(n <= msm_mem->capacity,
                    "zk_g2_msm_cached_async: n=%u exceeds msm_mem capacity=%u",
                    n, msm_mem->capacity);
-    PANIC_IF_FALSE(point_offset + n <= cached->n,
+    PANIC_IF_FALSE(static_cast<uint64_t>(point_offset) + n <= cached->n,
                    "zk_g2_msm_cached_async: point_offset=%u + n=%u exceeds cached points=%u",
                    point_offset, n, cached->n);
     PANIC_IF_FALSE(stream != nullptr, "zk_g2_msm_cached_async: stream is null");
@@ -562,7 +562,7 @@ void zk_g2_msm_cached_launch_async(
     PANIC_IF_FALSE(n <= msm_mem->capacity,
                    "zk_g2_msm_cached_launch_async: n=%u exceeds msm_mem capacity=%u",
                    n, msm_mem->capacity);
-    PANIC_IF_FALSE(point_offset + n <= cached->n,
+    PANIC_IF_FALSE(static_cast<uint64_t>(point_offset) + n <= cached->n,
                    "zk_g2_msm_cached_launch_async: point_offset=%u + n=%u exceeds cached points=%u",
                    point_offset, n, cached->n);
     PANIC_IF_FALSE(stream != nullptr, "zk_g2_msm_cached_launch_async: stream is null");
